% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDose-package.R
\docType{package}
\name{MetaDose-package}
\alias{MetaDose}
\alias{MetaDose-package}
\title{MetaDose: Dose–Response Meta-Regression for Meta-Analysis}
\description{
The `MetaDose` package provides tools for conducting linear and nonlinear
dose–response meta-regression using study-level summary data. It supports both
continuous and binary outcomes and allows modeling of dose–effect relationships
using linear trends or nonlinear restricted cubic splines.

The package is designed to facilitate transparent, flexible, and reproducible
dose–response meta-analyses, with built-in visualization of fitted
dose–response curves.
}
\seealso{
Useful functions:
\itemize{
  \item \code{\link{mdcont}} for dose–response meta-regression with continuous outcomes.
  \item \code{\link{mdbin}} for dose–response meta-regression with binary outcomes.
}
}
\author{
Ahmed Abdelmageed \email{ahmedelsaeedmassad@gmail.com}
}
\keyword{dose-response}
\keyword{meta-analysis}
\keyword{meta-regression}
\keyword{package}
